pyproject.toml
[project]
name = "aws-customer-intel-orchestrator"
version = "0.1.0"
description = "Strands-based multi-agent workflow for AWS customer intelligence."
requires-python = ">=3.10"

dependencies = [
  "strands-agents",
  "httpx",
  "pydantic",
  "strand>=0.1.8",
  "boto3",
  "fastapi>=0.127.0",
  "uvicorn>=0.30.0",
  "python-dotenv",
  "tavily-python>=0.7.0",
  "jinja2>=3.1.0",
  "a2a-sdk>=0.3.22",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "api*"]
exclude = ["reports"]

[project.scripts]
aws-intel-run = "api.main:run_server"

[tool.uv]
package = true
